move PersistentBrowser to new file
