+++ features/support/env.rb
@@ -16,19 +16,3 @@ end
 at_exit do
   Druid::PersistantBrowser.quit
 end
-
-module Druid
-  module PersistantBrowser
-    @@driver = false
-    def self.get_browser
-      if !@@driver
-        @@driver = Watir::Browser.new :firefox if ENV['DRIVER'] == 'WATIR'
-      end
-      @@driver
-    end
-
-    def self.quit
-      @@driver.quit
-    end
-  end
-end
